Python/week4_backend_ai/day26_python_ai/requirements.txt
# Day 26 - Python for AI Dependencies
numpy>=1.26.0
pandas>=2.1.0
openai>=1.6.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
tiktoken>=0.5.0
